ml_service/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
faiss-cpu
opencv-python
ultralytics
pillow
numpy
python-multipart